task(SDK-4214) - Adds logic for encryption migration based on new approach from the TAN
