+++ clevertap-core/src/main/java/com/clevertap/android/sdk/LocalDataStore.java
@@ -11,7 +11,6 @@ import androidx.annotation.RestrictTo.Scope;
 import androidx.annotation.WorkerThread;
 
 import com.clevertap.android.sdk.cryption.CryptHandler;
-import com.clevertap.android.sdk.cryption.CryptUtils;
 import com.clevertap.android.sdk.db.BaseDatabaseManager;
 import com.clevertap.android.sdk.db.DBAdapter;
 import com.clevertap.android.sdk.events.EventDetail;
@@ -575,7 +574,7 @@ public class LocalDataStore {
                         if (profile.get(piiKey) != null) {
                             Object value = profile.get(piiKey);
                             if (value instanceof String) {
-                                String encrypted = cryptHandler.encrypt((String) value, piiKey);
+                                String encrypted = cryptHandler.encrypt((String) value, piiKey, CryptHandler.EncryptionAlgorithm.AES_GCM);
                                 if (encrypted == null) {
                                     passFlag = false;
                                     continue;
@@ -587,8 +586,7 @@ public class LocalDataStore {
                     JSONObject jsonObjectEncrypted = new JSONObject(profile);
 
                     if (!passFlag)
-                        // todo replace with constant/ enum
-                        cryptHandler.updateEncryptionStateOnFailure(context, "currentStateDb");
+                        cryptHandler.updateMigrationFailureCount(context, false);
 
                     DBAdapter dbAdapter = baseDatabaseManager.loadDBAdapter(context);
                     long status = dbAdapter.storeUserProfile(profileID, deviceInfo.getDeviceID(), jsonObjectEncrypted);
