+++ clevertap-core/src/main/java/com/clevertap/android/sdk/cryption/AESGCMCrypt.kt
@@ -91,6 +91,7 @@ class AESGCMCrypt : Crypt() {
 
     private fun generateOrGetKey(): SecretKey? {
         return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            // TODO add support for API 21 and 22
             try {
                 val keyStore = KeyStore.getInstance("AndroidKeyStore")
                 keyStore.load(null)
