+++ clevertap-core/src/main/java/com/clevertap/android/sdk/login/LoginInfoProvider.java
@@ -59,12 +59,11 @@ public class LoginInfoProvider {
         if (isErrorDeviceId() || guid == null || key == null || identifier == null) {
             return;
         }
-        String encryptedIdentifier = cryptHandler.encrypt(identifier, key);
+        String encryptedIdentifier = cryptHandler.encrypt(identifier, key, CryptHandler.EncryptionAlgorithm.AES_GCM);
         if (encryptedIdentifier == null) {
             // If encrypted is null then fallback to plain text
             encryptedIdentifier = identifier;
-            // todo replace with constant/ enum
-            cryptHandler.updateEncryptionStateOnFailure(context, "currentStateCgk");
+            cryptHandler.updateMigrationFailureCount(context, false);
         }
         String cacheKey = key + "_" + encryptedIdentifier;
         JSONObject cache = getCachedGUIDs();
@@ -179,7 +178,7 @@ public class LoginInfoProvider {
         if (key == null || identifier == null) {
             return null;
         }
-        String encryptedIdentifier = cryptHandler.encrypt(identifier, key);
+        String encryptedIdentifier = cryptHandler.encrypt(identifier, key, CryptHandler.EncryptionAlgorithm.AES_GCM);
         String cacheKey = key + "_" + encryptedIdentifier;
         JSONObject cache = getCachedGUIDs();
         try {
